backend/requirements.txt
about-time==4.2.1
alive-progress==3.2.0
blinker==1.9.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.55.2
grapheme==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
mypy==1.13.0
mypy-extensions==1.0.0
numpy==2.2.0
optype==0.7.3
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.241126
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scipy==1.14.1
scipy-stubs==1.14.1.5
seaborn==0.13.2
six==1.17.0
types-click==7.1.8
types-Flask==1.1.6
types-Jinja2==2.11.9
types-MarkupSafe==1.1.10
types-pytz==2024.2.0.20241003
types-seaborn==0.13.2.20241204
types-Werkzeug==1.0.9
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
